pyproject.toml
[project]
name = "shop1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "coverage>=7.13.5",
    "django>=6.0.4",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.9",
    "drf-spectacular>=0.29.0",
    "python-slugify>=8.0.4",
]
